+++ connect/connect-core/connect-json/schema-registry/src/main/java/io/miragon/miranum/connect/json/registry/adapter/in/SchemaRestController.java
@@ -54,7 +54,7 @@ public class SchemaRestController {
     public ResponseEntity<JsonNode> getLatestSchema(@PathVariable final String ref) {
         try {
             final Schema schema = this.readSchemaUseCase.loadLatestSchema(ref);
-            return ResponseEntity.ok(schema.getJsonSchema());
+            return ResponseEntity.ok(schema.getJsonNode());
         } catch (final NoSuchElementException exception) {
             return ResponseEntity.notFound().build();
         }
