+++ connect/connect-core/connect-json/schema-registry/src/main/java/io/miragon/miranum/connect/json/registry/domain/Schema.java
@@ -11,6 +11,6 @@ public class Schema {
     private final String id;
     private final String ref;
     private final Integer version;
-    private final JsonNode jsonSchema;
+    private final JsonNode jsonNode;
 
 }
