+++ connect/connect-core/connect-json/schema-registry/src/test/resources/db_h2/V1__schema_registry.sql
@@ -5,7 +5,7 @@ create table miranum_schema_registry
             primary key,
     ref_          varchar(255)    not null,
     version_      int             not null,
-    json_schema_  text            not null
+    json_node_    text            not null
 );
 
 create index if not exists miranum_schema_registry_key_version_idx
